Rework the Main Tab and Workspace panel.
